refactor(evm): organize network implementations
Keep shared traits, aliases, and helpers in the EVM module root. Move network markers and halt-reason conversions into their network modules and consistently order EVM implementations and supporting code. Preserve public re-exports and execution behavior.
